requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# API Features
slowapi==0.1.9
python-multipart==0.0.6

# Frontend
streamlit==1.29.0
plotly==5.18.0
pygame==2.5.2

# Analytics
dbt-core==1.7.3
dbt-postgres==1.7.3
pandas==2.1.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Utilities
requests==2.31.0
python-dateutil==2.8.2
